# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: PicoClaw Serverless API
  version: 1.0.0
  description: |
    HTTP API for PicoClaw (serverless-oriented runtime).

    Notes:
    - All endpoints except `/health` require Bearer token authentication when `API_TOKEN` is set. When `API_TOKEN` is unset, authentication is disabled (auth-free mode for local/trusted-network deployments).
    - `POST /chat` supports both JSON response mode and SSE streaming mode.
    - Memory and conversation history are persisted via mounted storage and reused across separate requests.
    - Scheduled tasks are externally triggered by calling `/task/check`.
    - `POST /control/stop` can be used to request graceful save-and-exit.
    - Every response includes an `X-Request-ID` header for log correlation. If the caller sends an `X-Request-ID` header, the same value is echoed back; otherwise a new `req-<UUID>` is generated.
    - Every response includes `X-Build-Version` and `X-Build-Commit` headers for identifying the running build.
servers:
  - url: http://localhost:9000
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /task:
    post:
      tags:
        - Task
      summary: Create a scheduled task
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
            examples:
              once:
                value:
                  id: verify-task-once
                  prompt: 请只回复 ok
                  schedule_type: once
                  schedule_value: '2026-03-07T00:00:00'
                  context_mode: isolated
              interval:
                value:
                  prompt: 每 5 分钟输出一次状态
                  schedule_type: interval
                  schedule_value: '300000'
                  context_mode: isolated
              cron:
                value:
                  prompt: 生成每日摘要
                  schedule_type: cron
                  schedule_value: 0 9 * * 1-5
                  context_mode: group
                  conversation_id: conv-123
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid task payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /tasks:
    get:
      tags:
        - Task
      summary: List tasks
      operationId: listTasks
      responses:
        '200':
          description: Task list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /task/{task_id}:
    put:
      tags:
        - Task
      summary: Update a task
      operationId: updateTask
      parameters:
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
            examples:
              updateSchedule:
                value:
                  schedule_type: interval
                  schedule_value: '600000'
              pause:
                value:
                  status: paused
      responses:
        '200':
          description: Updated task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

    delete:
      tags:
        - Task
      summary: Delete a task
      operationId: deleteTask
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '204':
          description: Task deleted
        '400':
          description: Invalid path parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /task/trigger:
    post:
      tags:
        - Task
      summary: Trigger one specific task immediately
      operationId: triggerTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskTriggerRequest'
            examples:
              trigger:
                value:
                  task_id: verify-task-once
      responses:
        '200':
          description: Task execution result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskExecutionResult'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Task's conversation is currently processing another request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationBusyError'
        '500':
          description: Internal trigger failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /task/check:
    post:
      tags:
        - Task
      summary: Check due tasks and execute one
      description: |
        Fetches due tasks and executes only the first one.
        If no due tasks exist, returns checked=0.
      operationId: checkDueTasks
      responses:
        '200':
          description: No due task or one task execution completed
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/TaskCheckNoDueResponse'
                  - $ref: '#/components/schemas/TaskCheckDueResponse'
              examples:
                noDue:
                  value:
                    checked: 0
                    message: No due tasks
                dueExecuted:
                  value:
                    checked: 3
                    executed:
                      status: success
                      task_id: task-1
                      result: ok
                      duration_ms: 4874
                      next_run: null
                      error: null
                    remaining: 2
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          description: Unexpected runtime failure while checking/executing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskCheckErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API_TOKEN
  parameters:
...
    TaskId:
      name: task_id
      in: path
      required: true
      description: Task ID
      schema:
        type: string
        minLength: 1
  headers:
    X-Request-ID:
      description: |
        Request tracking identifier for log correlation.
        Echoes the caller-provided `X-Request-ID` header, or a generated `req-<UUID>` if none was sent.
        Present on every response, including error and unauthenticated responses.
      schema:
        type: string
        example: req-a1b2c3d4-e5f6-7890-abcd-ef1234567890
    X-Build-Version:
      description: |
        Application version from `package.json`, injected at Docker build time.
        Present on every response.
      schema:
        type: string
        example: 1.2.16
    X-Build-Commit:
      description: |
        Short git commit hash of the build, injected at Docker build time.
        Present on every response.
      schema:
        type: string
        example: abc1234
  responses:
    Unauthorized:
      description: Unauthorized (missing or invalid bearer token)
      headers:
        X-Request-ID:
          $ref: '#/components/headers/X-Request-ID'
        X-Build-Version:
          $ref: '#/components/headers/X-Build-Version'
        X-Build-Commit:
          $ref: '#/components/headers/X-Build-Commit'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          examples:
            unauthorized:
              value:
                error: Unauthorized
  schemas:
    ApiError:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          type: string
...
    Task:
      type: object
      additionalProperties: false
      required:
        - id
        - conversation_id
        - prompt
        - schedule_type
        - schedule_value
        - context_mode
        - next_run
        - status
        - created_at
      properties:
        id:
          type: string
        conversation_id:
          type: string
        prompt:
          type: string
        schedule_type:
          type: string
          enum:
            - cron
            - interval
            - once
        schedule_value:
          type: string
        context_mode:
          type: string
          enum:
            - group
            - isolated
        next_run:
          type: string
          format: date-time
          nullable: true
        last_run:
          type: string
          format: date-time
          nullable: true
        last_result:
          type: string
          nullable: true
        status:
          type: string
          enum:
            - active
            - paused
            - completed
        created_at:
          type: string
          format: date-time
    CreateTaskRequest:
      type: object
      additionalProperties: false
      required:
        - prompt
        - schedule_type
        - schedule_value
      properties:
        id:
          type: string
          description: Optional custom task ID.
        prompt:
          type: string
          minLength: 1
        schedule_type:
          type: string
          enum:
            - cron
            - interval
            - once
        schedule_value:
          type: string
          minLength: 1
          description: |
            - cron: cron expression (TZ-aware)
            - interval: milliseconds string (positive integer)
            - once: local timestamp string without timezone suffix
        context_mode:
          type: string
          enum:
            - group
            - isolated
          default: isolated
        conversation_id:
          type: string
          description: Existing conversation ID. Auto-created if omitted.
    UpdateTaskRequest:
      type: object
      additionalProperties: false
      properties:
        prompt:
          type: string
          minLength: 1
        schedule_type:
          type: string
          enum:
            - cron
            - interval
            - once
        schedule_value:
          type: string
          minLength: 1
        context_mode:
          type: string
          enum:
            - group
            - isolated
        status:
          type: string
          enum:
            - active
            - paused
            - completed
        conversation_id:
          type: string
    TaskListResponse:
      type: object
      additionalProperties: false
      required:
        - tasks
      properties:
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/Task'
    TaskTriggerRequest:
      type: object
      additionalProperties: false
      required:
        - task_id
      properties:
        task_id:
          type: string
          minLength: 1
    TaskExecutionResult:
      type: object
      additionalProperties: false
      required:
        - status
        - task_id
        - result
        - duration_ms
        - next_run
      properties:
        status:
          type: string
          enum:
            - success
            - timeout
            - error
        task_id:
          type: string
        result:
          type: string
          nullable: true
        duration_ms:
          type: integer
          minimum: 0
        next_run:
          type: string
          format: date-time
          nullable: true
        error:
          type: string
          nullable: true
    TaskCheckNoDueResponse:
      type: object
      additionalProperties: false
      required:
        - checked
        - message
      properties:
        checked:
          type: integer
          enum:
            - 0
        message:
          type: string
          enum:
            - No due tasks
    TaskCheckDueResponse:
      type: object
      additionalProperties: false
      required:
        - checked
        - executed
        - remaining
      properties:
        checked:
          type: integer
          minimum: 1
        executed:
          $ref: '#/components/schemas/TaskExecutionResult'
        remaining:
          type: integer
          minimum: 0
    TaskCheckErrorExecuted:
      type: object
      additionalProperties: false
      required:
        - task_id
        - status
        - error
      properties:
        task_id:
          type: string
        status:
          type: string
          enum:
            - error
        error:
          type: string
    TaskCheckErrorResponse:
      type: object
      additionalProperties: false
      required:
        - checked
        - executed
        - remaining
      properties:
        checked:
          type: integer
          minimum: 1
        executed:
          $ref: '#/components/schemas/TaskCheckErrorExecuted'
        remaining:
          type: integer
          minimum: 0
...
    ConversationBusyError:
      type: object
      additionalProperties: false
      required:
        - error
        - conversation_id
      properties:
        error:
          type: string
        conversation_id:
          type: string
...